+++ returns/context/requires_context.py
@@ -211,7 +211,6 @@ class RequiresContext(
 
         See also:
             - https://wiki.haskell.org/Lifting
-            - https://github.com/witchcrafters/witchcraft
             - https://en.wikipedia.org/wiki/Natural_transformation
 
         """