+++ returns/context/requires_context_io_result.py
@@ -575,7 +575,6 @@ class RequiresContextIOResult(
 
         See also:
             - https://wiki.haskell.org/Lifting
-            - https://github.com/witchcrafters/witchcraft
             - https://en.wikipedia.org/wiki/Natural_transformation
 
         """