+++ returns/context/requires_context_result.py
@@ -468,7 +468,6 @@ class RequiresContextResult(
 
         See also:
             - https://wiki.haskell.org/Lifting
-            - https://github.com/witchcrafters/witchcraft
             - https://en.wikipedia.org/wiki/Natural_transformation
 
         """